`get_ssl_certificate_id()` checks more thoroughly.
Updated the `get_ssl_certificate_id()` function to check that `SSL_CLIENT_VERIFY` is set to `SUCCESS` before hashing the certificate values.

This is only useful if the web server (Apache, Nginx, etc.) is set to **optionally** allow client certificate authentication. (If it's required, the server itself will reject the connection.)